--- parser/js/src/main/scala/jawn/SupportParserPlatform.scala
@@ -1,3 +1,0 @@
-package org.typelevel.jawn
-
-private[jawn] trait SupportParserPlatform[J]
