+++ parser/jvm/src/main/scala/jawn/SupportParser.scala
@@ -1,9 +1,12 @@
 package org.typelevel.jawn
 
+import java.io.File
 import java.nio.ByteBuffer
+import java.nio.channels.ReadableByteChannel
 import scala.util.Try
 
-trait SupportParser[J] extends SupportParserPlatform[J] {
+// Had to be duplicated in whole (instead of "platformed") for bincompat
+trait SupportParser[J] {
   implicit def facade: Facade[J]
 
   def parseUnsafe(s: String): J =
@@ -14,6 +17,15 @@ trait SupportParser[J] extends SupportParserPlatform[J] {
 
   def parseFromCharSequence(cs: CharSequence): Try[J] =
     Try(new CharSequenceParser[J](cs).parse())
+
+  def parseFromPath(path: String): Try[J] =
+    Try(ChannelParser.fromFile[J](new File(path)).parse())
+
+  def parseFromFile(file: File): Try[J] =
+    Try(ChannelParser.fromFile[J](file).parse())
+
+  def parseFromChannel(ch: ReadableByteChannel): Try[J] =
+    Try(ChannelParser.fromChannel[J](ch).parse())
 
   def parseFromByteBuffer(buf: ByteBuffer): Try[J] =
     Try(new ByteBufferParser[J](buf).parse())
